Upgrade-13/12/2019

+++ ForaTeknoloji.PresentationLayer/Controllers/BlockController.cs
@@ -1,6 +1,7 @@
 ﻿using ForaTeknoloji.BusinessLayer.Abstract;
 using ForaTeknoloji.Entities.Entities;
 using ForaTeknoloji.PresentationLayer.Filters;
+using ForaTeknoloji.PresentationLayer.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -14,9 +15,19 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
     public class BlockController : Controller
     {
         private IBloklarService _bloklarService;
-        public BlockController(IBloklarService bloklarService)
+        private IDBUsersService _dBUsersService;
+        public DBUsers user;
+        public DBUsers permissionUser;
+        public BlockController(IBloklarService bloklarService, IDBUsersService dBUsersService)
         {
+            user = CurrentSession.User;
+            if (user == null)
+            {
+                user = new DBUsers();
+            }
             _bloklarService = bloklarService;
+            _dBUsersService = dBUsersService;
+            permissionUser = _dBUsersService.GetAllDBUsers().Find(x => x.Kullanici_Adi == user.Kullanici_Adi);
         }
 
 
@@ -31,34 +42,50 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
         [HttpPost]
         public ActionResult Create(Bloklar bloklar)
         {
-            if (ModelState.IsValid)
+            if (permissionUser.SysAdmin == false)
             {
-                if (bloklar.Adi != null)
-                {
-                    var ID = _bloklarService.GetAllBloklar().Count;
-                    if (ID == 0)
-                        _bloklarService.DeleteAll();
-
-                    _bloklarService.AddBloklar(bloklar);
-                    return RedirectToAction("Index");
-                }
-                throw new Exception("Yanlış yada eksik karakter girdiniz");
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return RedirectToAction("Index");
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    if (bloklar.Adi != null)
+                    {
+                        var ID = _bloklarService.GetAllBloklar().Count;
+                        if (ID == 0)
+                            _bloklarService.DeleteAll();
+
+                        _bloklarService.AddBloklar(bloklar);
+                        return RedirectToAction("Index");
+                    }
+                    throw new Exception("Yanlış yada eksik karakter girdiniz");
+                }
+                return RedirectToAction("Index");
+            }
+
+
         }
 
         public ActionResult Delete(int id = -1)
         {
-            if (id != -1)
+            if (permissionUser.SysAdmin == false)
             {
-                Bloklar bloklar = _bloklarService.GetById(id);
-                if (bloklar != null)
-                {
-                    _bloklarService.DeleteBloklar(bloklar);
-                    return Json(true, JsonRequestBehavior.AllowGet);
-                }
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return Json(false, JsonRequestBehavior.AllowGet);
+            else
+            {
+                if (id != -1)
+                {
+                    Bloklar bloklar = _bloklarService.GetById(id);
+                    if (bloklar != null)
+                    {
+                        _bloklarService.DeleteBloklar(bloklar);
+                        return Json(true, JsonRequestBehavior.AllowGet);
+                    }
+                }
+                return Json(false, JsonRequestBehavior.AllowGet);
+            }
         }
 
         public ActionResult Edit(int? id)
@@ -73,22 +100,31 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
                 return HttpNotFound();
             }
             return View(bloklar);
+
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Bloklar bloklar)
         {
-            if (ModelState.IsValid)
+            if (permissionUser.SysAdmin == false)
             {
-                var blok = _bloklarService.GetById(bloklar.Blok_No);
-                if (blok != null)
-                {
-                    _bloklarService.UpdateBloklar(bloklar);
-                    return RedirectToAction("Index");
-                }
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return View(bloklar);
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    var blok = _bloklarService.GetById(bloklar.Blok_No);
+                    if (blok != null)
+                    {
+                        _bloklarService.UpdateBloklar(bloklar);
+                        return RedirectToAction("Index");
+                    }
+                }
+                return View(bloklar);
+            }
+
         }
