+++ ForaTeknoloji.PresentationLayer/Controllers/CompanyController.cs
@@ -1,6 +1,7 @@
 ﻿using ForaTeknoloji.BusinessLayer.Abstract;
 using ForaTeknoloji.Entities.Entities;
 using ForaTeknoloji.PresentationLayer.Filters;
+using ForaTeknoloji.PresentationLayer.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,9 +16,19 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
     public class CompanyController : Controller
     {
         private ISirketService _sirketService;
-        public CompanyController(ISirketService sirketService)
+        private IDBUsersService _dBUsersService;
+        public DBUsers user;
+        public DBUsers permissionUser;
+        public CompanyController(ISirketService sirketService, IDBUsersService dBUsersService)
         {
+            user = CurrentSession.User;
+            if (user == null)
+            {
+                user = new DBUsers();
+            }
             _sirketService = sirketService;
+            _dBUsersService = dBUsersService;
+            permissionUser = _dBUsersService.GetAllDBUsers().Find(x => x.Kullanici_Adi == user.Kullanici_Adi);
         }
 
 
@@ -31,34 +42,50 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
         [HttpPost]
         public ActionResult Create(Sirketler Sirket)
         {
-            if (ModelState.IsValid)
+            if (permissionUser.SysAdmin == false)
             {
-                if (Sirket.Adi != null)
-                {
-                    var ID = _sirketService.GetAllSirketler().Count;
-                    if (ID == 0)
-                        _sirketService.DeleteAll();
-
-                    _sirketService.AddSirket(Sirket);
-                    return RedirectToAction("Index");
-                }
-                throw new Exception("Yanlış yada eksik karakter girdiniz.");
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return RedirectToAction("Index");
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    if (Sirket.Adi != null)
+                    {
+                        var ID = _sirketService.GetAllSirketler().Count;
+                        if (ID == 0)
+                            _sirketService.DeleteAll();
+
+                        _sirketService.AddSirket(Sirket);
+                        return RedirectToAction("Index");
+                    }
+                    throw new Exception("Yanlış yada eksik karakter girdiniz.");
+                }
+                return RedirectToAction("Index");
+            }
+
+
         }
 
         public ActionResult Delete(int id = -1)
         {
-            if (id != -1)
+            if (permissionUser.SysAdmin == false)
             {
-                Sirketler sirket = _sirketService.GetById(id);
-                if (sirket != null)
-                {
-                    _sirketService.DeleteSirket(sirket);
-                    return Json(true, JsonRequestBehavior.AllowGet);
-                }
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return Json(false, JsonRequestBehavior.AllowGet);
+            else
+            {
+                if (id != -1)
+                {
+                    Sirketler sirket = _sirketService.GetById(id);
+                    if (sirket != null)
+                    {
+                        _sirketService.DeleteSirket(sirket);
+                        return Json(true, JsonRequestBehavior.AllowGet);
+                    }
+                }
+                return Json(false, JsonRequestBehavior.AllowGet);
+            }
         }
 
         public ActionResult Edit(int? id)
@@ -79,16 +106,23 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Sirketler sirketler)
         {
-            if (ModelState.IsValid)
+            if (permissionUser.SysAdmin == false)
             {
-                var sirket = _sirketService.GetById(sirketler.Sirket_No);
-                if (sirket != null)
-                {
-                    _sirketService.UpdateSirket(sirketler);
-                    return RedirectToAction("Index");
-                }
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return View(sirketler);
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    var sirket = _sirketService.GetById(sirketler.Sirket_No);
+                    if (sirket != null)
+                    {
+                        _sirketService.UpdateSirket(sirketler);
+                        return RedirectToAction("Index");
+                    }
+                }
+                return View(sirketler);
+            }
         }
 
     }
