+++ ForaTeknoloji.PresentationLayer/Controllers/EmploymentController.cs
@@ -1,6 +1,7 @@
 ﻿using ForaTeknoloji.BusinessLayer.Abstract;
 using ForaTeknoloji.Entities.Entities;
 using ForaTeknoloji.PresentationLayer.Filters;
+using ForaTeknoloji.PresentationLayer.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -14,9 +15,19 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
     public class EmploymentController : Controller
     {
         private IGorevlerService _gorevlerService;
-        public EmploymentController(IGorevlerService gorevlerService)
+        private IDBUsersService _dBUsersService;
+        public DBUsers user;
+        public DBUsers permissionUser;
+        public EmploymentController(IGorevlerService gorevlerService, IDBUsersService dBUsersService)
         {
+            user = CurrentSession.User;
+            if (user == null)
+            {
+                user = new DBUsers();
+            }
             _gorevlerService = gorevlerService;
+            _dBUsersService = dBUsersService;
+            permissionUser = _dBUsersService.GetAllDBUsers().Find(x => x.Kullanici_Adi == user.Kullanici_Adi);
         }
 
 
@@ -29,34 +40,48 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
         [HttpPost]
         public ActionResult Create(Gorevler Gorev)
         {
-            if (ModelState.IsValid)
+            if (permissionUser.SysAdmin == false)
             {
-                if (Gorev.Adi != null)
-                {
-                    var ID = _gorevlerService.GetAllGorevler().Count;
-                    if (ID == 0)
-                        _gorevlerService.DeleteAll();
-
-                    _gorevlerService.AddGorev(Gorev);
-                    return RedirectToAction("Index");
-                }
-                throw new Exception("Yanlış yada eksik karakter girdiniz.");
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return RedirectToAction("Index");
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    if (Gorev.Adi != null)
+                    {
+                        var ID = _gorevlerService.GetAllGorevler().Count;
+                        if (ID == 0)
+                            _gorevlerService.DeleteAll();
+
+                        _gorevlerService.AddGorev(Gorev);
+                        return RedirectToAction("Index");
+                    }
+                    throw new Exception("Yanlış yada eksik karakter girdiniz.");
+                }
+                return RedirectToAction("Index");
+            }
         }
 
         public ActionResult Delete(int id = -1)
         {
-            if (id != -1)
+            if (permissionUser.SysAdmin == false)
             {
-                Gorevler gorev = _gorevlerService.GetById(id);
-                if (gorev != null)
-                {
-                    _gorevlerService.DeleteGorev(gorev);
-                    return Json(true, JsonRequestBehavior.AllowGet);
-                }
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return Json(false, JsonRequestBehavior.AllowGet);
+            else
+            {
+                if (id != -1)
+                {
+                    Gorevler gorev = _gorevlerService.GetById(id);
+                    if (gorev != null)
+                    {
+                        _gorevlerService.DeleteGorev(gorev);
+                        return Json(true, JsonRequestBehavior.AllowGet);
+                    }
+                }
+                return Json(false, JsonRequestBehavior.AllowGet);
+            }
         }
 
         public ActionResult Edit(int? id)
@@ -78,16 +103,24 @@ namespace ForaTeknoloji.PresentationLayer.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit(Gorevler gorevler)
         {
-            if (ModelState.IsValid)
+            if (permissionUser.SysAdmin == false)
             {
-                var gorev = _gorevlerService.GetById(gorevler.Gorev_No);
-                if (gorev != null)
-                {
-                    _gorevlerService.UpdateGorev(gorevler);
-                    return RedirectToAction("Index");
-                }
+                throw new Exception("Yetkisiz Erişim!");
             }
-            return View(gorevler);
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    var gorev = _gorevlerService.GetById(gorevler.Gorev_No);
+                    if (gorev != null)
+                    {
+                        _gorevlerService.UpdateGorev(gorevler);
+                        return RedirectToAction("Index");
+                    }
+                }
+                return View(gorevler);
+            }
+
         }
